Moving classes to files
